backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
SQLAlchemy>=2.0.0,<3.0.0
asyncpg>=0.30.0,<1.0.0
aiosqlite>=0.21.0,<1.0.0
pydantic-settings>=2.8.0,<3.0.0
python-multipart>=0.0.20,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt==4.0.1
pytest>=8.0.0,<9.0.0
